refactoring all code, adding features
Now it uses a custom hook to control the global state to avoid fetch data that has been fetched previously; now it have all information we can get from API, and everything is wired so now it possible to navigate to any data of the API easily.

TODO: the search bar doesnt work in details pages, also lack of styles in details pages.